Added workaround to prevent issues with utf8
Without this workaround you would see 'strange' characters appearing
when using the plugin with utf-8 encoded data. This is apparently an
issue (bug?) with PHP's DOM extension which defaults to using Latin1
when loading HTML without a metatag describing the encoding. Explicitly
setting the encoding does not work. This workaround however should solve
all this. It ain't pretty....