Suppress malformed html warnings in the logs
Using the libxml_use_internal_errors(true) function to suppress errors. This
prevents the logs from becoming flooded with warnings of mal-formed
html. Since most of the time we have no control over the html these
warnings are useless and clutter our logs. In case we need to access the
errors simply set the to libxml_use_internal_errors(false)