Link contains visual element than don't add icons
Added code which prevents an icon to clash with another visual element:

If a html a tag contains an image/video/object or embed html tag it will
not add an icon to this link. This prevents the unwanted side-effect of
having both an icon and another visual element clash with each other

This bumps the version of this plugin up to 3.1 It should not cause any
backwards compatibility issues from version 3.0 upwards. Versions lower
than 3.0 are NOT BACKWARDS COMPATIBLE!